Improve Keyboard placement
Improved the Keyboard placement by to attaching it to the word_view.
This puts the keyboard underneath the word_view so it doesn't move
around when the keys are removed by being pressed.

To make the keyboard fit better the word_view's height was lowered and
the text centered. Also changed the font to monospace so all characters
use the same amount of space. This resolves the jarring effect of the
word_view moving around when a letter is successfully guessed.