Add IXD features around winning/losing
Added playing of loss and win sound when you win or losses the round.
You will also get displayed the correct word when you lose written in
red. If you win and guess the word the word is written in green. There
is also a nice delay before the next round is started giving the user
time to see if he managed to correctly guess the word or not and listen
to the sound effects. When the session is over the user will get an alert
dialog asking if the game should restart or exit.