Add getting words from wordlist.csv
The WordFileDAO now gets the words used by the game from a wordlist.
Since the wordlist is in resources the DAO needs a context. GameManager
so providing a context to GameManager is difficult for this particular
implementation of GameManager. Therefore the App class provides a context
through a static method which is a hacky solution at best.